+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 
 ## [Unreleased](https://github.com/techprimate/TPPDF/tree/HEAD) (2020-??-??)
-[Full Changelog](https://github.com/techprimate/TPPDF/compare/2.1.0...HEAD)
+[Full Changelog](https://github.com/techprimate/TPPDF/compare/2.1.1...HEAD)
 
 **Implemented enhancements:**
 
@@ -11,7 +11,19 @@
 
 **Merged pull requests:**
 
-## [2.1.0](https://github.com/techprimate/TPPDF/tree/HEAD) (2020-06-15)
+## [2.1.1](https://github.com/techprimate/TPPDF/tree/2.1.1) (2020-06-16)
+[Full Changelog](https://github.com/techprimate/TPPDF/compare/2.1.0...2.1.1)
+
+**Implemented enhancements:**
+
+- Added raw representable to `PDFTableCellAlignment`
+- Added support for groups inside section columns
+
+**Fixed bugs:**
+
+- Fixed indentations inside sections
+
+## [2.1.0](https://github.com/techprimate/TPPDF/tree/2.1.0) (2020-06-15)
 [Full Changelog](https://github.com/techprimate/TPPDF/compare/2.0.1...2.1.0)
 
 **Implemented enhancements:**